refactor: refactoring unit testing
