+++ test/NUMBER.test.js
@@ -1,11 +1,11 @@
 import { describe, assert, expect, test, beforeEach, afterEach, beforeAll, vi } from 'vitest'
-import { flushPromises, config } from '@vue/test-utils'
+import { mount, flushPromises, enableAutoUnmount, config } from '@vue/test-utils'
 import optionData from '@ER-server/routes/Filter/data/options.js'
 import * as ElementPlusIconsVue from '@element-plus/icons-vue'
 import NAME from '@ER/filter/name.js'
 import utils from '@ER/utils'
 import { nextTick } from 'vue'
-import { getSelectOptions, _mount } from '../utils.js'
+import { getSelectOptions, _mount } from './utils.js'
 config.global.components.Delete = ElementPlusIconsVue.Delete
 const getOptions = async () => {
   return new Promise((resolve, reject) => {
@@ -14,10 +14,10 @@ const getOptions = async () => {
     })
   })
 }
-describe('renderType: TEXT', () => {
+describe('renderType: NUMBER', () => {
   let wrapper = {}
   const handleListener = vi.fn()
-  beforeAll(() => {
+  beforeAll(async () => {
     wrapper = _mount(`
       <everright-filter
         :getOptions="getOptions"
@@ -37,72 +37,62 @@ describe('renderType: TEXT', () => {
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find('.Everright-filter-TriggerComponent').trigger('click')
     document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-node')[0].click()
     await nextTick()
-    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[5].click()
+    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[6].click()
     await flushPromises()
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     expect(getSelectOptions(utils.getTestId('operator-popperClass', 'id')).map(e => utils.getAttrs(e).value)).toEqual([
       'equal',
-      'one_of',
       'not_equal',
-      'contains',
-      'not_contain',
+      'greater_than',
+      'greater_than_equal',
+      'less_than',
+      'less_than_equal',
+      'between',
       'empty',
       'not_empty'
     ])
-    let inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-input`, 'id'))
+    let inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
     const inputWrapper = inputElm.find('input')
     const nativeInput = inputWrapper.element
     expect(inputElm.exists()).toBe(true)
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toEqual({})
     await new Promise(resolve => setTimeout(resolve, 100))
-    inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-input`, 'id'))
+    inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
     expect(inputElm.classes()).toContain('ERFILTER-ERROR')
     nativeInput.value = '1'
     await inputWrapper.trigger('input')
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
-  test('operator: "style=one_of" && value is not empty', async () => {
+  test('operator: "style=range" && value is not empty', async () => {
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find('.Everright-filter-TriggerComponent').trigger('click')
     document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-node')[0].click()
     await nextTick()
-    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[5].click()
+    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[6].click()
     await flushPromises()
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    expect(selectOptions.map(e => utils.getAttrs(e).value)).toEqual([
-      'equal',
-      'one_of',
-      'not_equal',
-      'contains',
-      'not_contain',
-      'empty',
-      'not_empty'
-    ])
-    selectOptions[1].click()
-    await nextTick()
-    let selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
+    selectOptions[6].click()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toEqual({})
     await new Promise(resolve => setTimeout(resolve, 100))
-    selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
-    expect(selectElm.classes()).toContain('ERFILTER-ERROR')
-    const inputWrapper = selectElm.find('.el-select__input')
-    const selectInputEl = inputWrapper.element
-    selectInputEl.value = '1'
-    await new Promise(resolve => setTimeout(resolve, 100))
-    await inputWrapper.trigger('input')
-    getSelectOptions(utils.getTestId(`${NAME.TEXTTYPE}-select-popperClass`, 'id'))[0].click()
-    await nextTick()
+    const number0Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
+    const number1Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number1`, 'id'))
+    expect(number0Elm.classes()).toContain('ERFILTER-ERROR')
+    expect(number1Elm.classes()).toContain('ERFILTER-ERROR')
+    number0Elm.find('input').element.value = '10'
+    await number0Elm.find('input').trigger('input')
+    number1Elm.find('input').element.value = '20'
+    await number1Elm.find('input').trigger('input')
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
   test('operator: "style=none"', async () => {
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find('.Everright-filter-TriggerComponent').trigger('click')
     document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-node')[0].click()
     await nextTick()
-    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[5].click()
+    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[6].click()
     await flushPromises()
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    selectOptions[5].click()
+    selectOptions[7].click()
     await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
@@ -112,8 +102,8 @@ describe('renderType: TEXT', () => {
         conditions: [
           {
             operator: 'equal',
-            property: 'text',
-            value: 'hello!'
+            property: 'rating',
+            value: 2000
           }
         ],
         logicalOperator: 'and'
@@ -128,13 +118,17 @@ describe('renderType: TEXT', () => {
     await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
-  test('Modifying the operatorStyle from none to tags', async () => {
+  test('Modifying the operatorStyle from range to none', async () => {
     wrapper.findComponent({ ref: 'ERfilterRef' }).vm.setData({
       filters: [{
         conditions: [
           {
-            operator: 'empty',
-            property: 'text'
+            operator: 'between',
+            property: 'rating',
+            value: [
+              1,
+              123
+            ]
           }
         ],
         logicalOperator: 'and'
@@ -144,29 +138,38 @@ describe('renderType: TEXT', () => {
     await new Promise(resolve => setTimeout(resolve, 100))
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    expect(selectOptions.map(e => utils.getAttrs(e).value)).toEqual([
-      'equal',
-      'one_of',
-      'not_equal',
-      'contains',
-      'not_contain',
-      'empty',
-      'not_empty'
-    ])
-    selectOptions[1].click()
+    selectOptions[7].click()
     await nextTick()
-    let selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
+    expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
+  })
+  test('Modifying the operatorStyle from none to range', async () => {
+    wrapper.findComponent({ ref: 'ERfilterRef' }).vm.setData({
+      filters: [{
+        conditions: [
+          {
+            operator: 'empty',
+            property: 'rating'
+          }
+        ],
+        logicalOperator: 'and'
+      }],
+      logicalOperator: 'and'
+    })
+    await new Promise(resolve => setTimeout(resolve, 100))
+    await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
+    const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
+    selectOptions[6].click()
+    await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toEqual({})
     await new Promise(resolve => setTimeout(resolve, 100))
-    selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
-    expect(selectElm.classes()).toContain('ERFILTER-ERROR')
-    const inputWrapper = selectElm.find('.el-select__input')
-    const selectInputEl = inputWrapper.element
-    selectInputEl.value = '1'
-    await new Promise(resolve => setTimeout(resolve, 100))
-    await inputWrapper.trigger('input')
-    getSelectOptions(utils.getTestId(`${NAME.TEXTTYPE}-select-popperClass`, 'id'))[0].click()
-    await nextTick()
+    const number0Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
+    const number1Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number1`, 'id'))
+    expect(number0Elm.classes()).toContain('ERFILTER-ERROR')
+    expect(number1Elm.classes()).toContain('ERFILTER-ERROR')
+    number0Elm.find('input').element.value = '10'
+    await number0Elm.find('input').trigger('input')
+    number1Elm.find('input').element.value = '20'
+    await number1Elm.find('input').trigger('input')
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
 })
