+++ test/TEXT.test.js
@@ -1,11 +1,11 @@
 import { describe, assert, expect, test, beforeEach, afterEach, beforeAll, vi } from 'vitest'
-import { mount, flushPromises, enableAutoUnmount, config } from '@vue/test-utils'
+import { flushPromises, config } from '@vue/test-utils'
 import optionData from '@ER-server/routes/Filter/data/options.js'
 import * as ElementPlusIconsVue from '@element-plus/icons-vue'
 import NAME from '@ER/filter/name.js'
 import utils from '@ER/utils'
 import { nextTick } from 'vue'
-import { getSelectOptions, _mount } from '../utils.js'
+import { getSelectOptions, _mount } from './utils.js'
 config.global.components.Delete = ElementPlusIconsVue.Delete
 const getOptions = async () => {
   return new Promise((resolve, reject) => {
@@ -14,10 +14,10 @@ const getOptions = async () => {
     })
   })
 }
-describe('renderType: NUMBER', () => {
+describe('renderType: TEXT', () => {
   let wrapper = {}
   const handleListener = vi.fn()
-  beforeAll(async () => {
+  beforeAll(() => {
     wrapper = _mount(`
       <everright-filter
         :getOptions="getOptions"
@@ -37,62 +37,72 @@ describe('renderType: NUMBER', () => {
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find('.Everright-filter-TriggerComponent').trigger('click')
     document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-node')[0].click()
     await nextTick()
-    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[6].click()
+    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[5].click()
     await flushPromises()
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     expect(getSelectOptions(utils.getTestId('operator-popperClass', 'id')).map(e => utils.getAttrs(e).value)).toEqual([
       'equal',
+      'one_of',
       'not_equal',
-      'greater_than',
-      'greater_than_equal',
-      'less_than',
-      'less_than_equal',
-      'between',
+      'contains',
+      'not_contain',
       'empty',
       'not_empty'
     ])
-    let inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
+    let inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-input`, 'id'))
     const inputWrapper = inputElm.find('input')
     const nativeInput = inputWrapper.element
     expect(inputElm.exists()).toBe(true)
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toEqual({})
     await new Promise(resolve => setTimeout(resolve, 100))
-    inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
+    inputElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-input`, 'id'))
     expect(inputElm.classes()).toContain('ERFILTER-ERROR')
     nativeInput.value = '1'
     await inputWrapper.trigger('input')
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
-  test('operator: "style=range" && value is not empty', async () => {
+  test('operator: "style=one_of" && value is not empty', async () => {
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find('.Everright-filter-TriggerComponent').trigger('click')
     document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-node')[0].click()
     await nextTick()
-    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[6].click()
+    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[5].click()
     await flushPromises()
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    selectOptions[6].click()
+    expect(selectOptions.map(e => utils.getAttrs(e).value)).toEqual([
+      'equal',
+      'one_of',
+      'not_equal',
+      'contains',
+      'not_contain',
+      'empty',
+      'not_empty'
+    ])
+    selectOptions[1].click()
+    await nextTick()
+    let selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toEqual({})
     await new Promise(resolve => setTimeout(resolve, 100))
-    const number0Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
-    const number1Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number1`, 'id'))
-    expect(number0Elm.classes()).toContain('ERFILTER-ERROR')
-    expect(number1Elm.classes()).toContain('ERFILTER-ERROR')
-    number0Elm.find('input').element.value = '10'
-    await number0Elm.find('input').trigger('input')
-    number1Elm.find('input').element.value = '20'
-    await number1Elm.find('input').trigger('input')
+    selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
+    expect(selectElm.classes()).toContain('ERFILTER-ERROR')
+    const inputWrapper = selectElm.find('.el-select__input')
+    const selectInputEl = inputWrapper.element
+    selectInputEl.value = '1'
+    await new Promise(resolve => setTimeout(resolve, 100))
+    await inputWrapper.trigger('input')
+    getSelectOptions(utils.getTestId(`${NAME.TEXTTYPE}-select-popperClass`, 'id'))[0].click()
+    await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
   test('operator: "style=none"', async () => {
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find('.Everright-filter-TriggerComponent').trigger('click')
     document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-node')[0].click()
     await nextTick()
-    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[6].click()
+    document.querySelector(utils.getTestId(`${NAME.TRIGGERCOMPONENT}-popperClass`, 'id')).querySelectorAll('.el-cascader-menu')[1].querySelectorAll('.el-cascader-node')[5].click()
     await flushPromises()
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    selectOptions[7].click()
+    selectOptions[5].click()
     await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
@@ -102,8 +112,8 @@ describe('renderType: NUMBER', () => {
         conditions: [
           {
             operator: 'equal',
-            property: 'rating',
-            value: 2000
+            property: 'text',
+            value: 'hello!'
           }
         ],
         logicalOperator: 'and'
@@ -118,37 +128,13 @@ describe('renderType: NUMBER', () => {
     await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
-  test('Modifying the operatorStyle from range to none', async () => {
-    wrapper.findComponent({ ref: 'ERfilterRef' }).vm.setData({
-      filters: [{
-        conditions: [
-          {
-            operator: 'between',
-            property: 'rating',
-            value: [
-              1,
-              123
-            ]
-          }
-        ],
-        logicalOperator: 'and'
-      }],
-      logicalOperator: 'and'
-    })
-    await new Promise(resolve => setTimeout(resolve, 100))
-    await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
-    const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    selectOptions[7].click()
-    await nextTick()
-    expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
-  })
-  test('Modifying the operatorStyle from none to range', async () => {
+  test('Modifying the operatorStyle from none to tags', async () => {
     wrapper.findComponent({ ref: 'ERfilterRef' }).vm.setData({
       filters: [{
         conditions: [
           {
             operator: 'empty',
-            property: 'rating'
+            property: 'text'
           }
         ],
         logicalOperator: 'and'
@@ -158,18 +144,29 @@ describe('renderType: NUMBER', () => {
     await new Promise(resolve => setTimeout(resolve, 100))
     await wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId('operator', 'id')).trigger('click')
     const selectOptions = getSelectOptions(utils.getTestId('operator-popperClass', 'id'))
-    selectOptions[6].click()
+    expect(selectOptions.map(e => utils.getAttrs(e).value)).toEqual([
+      'equal',
+      'one_of',
+      'not_equal',
+      'contains',
+      'not_contain',
+      'empty',
+      'not_empty'
+    ])
+    selectOptions[1].click()
     await nextTick()
+    let selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toEqual({})
     await new Promise(resolve => setTimeout(resolve, 100))
-    const number0Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number0`, 'id'))
-    const number1Elm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.NUMBERTYPE}-number1`, 'id'))
-    expect(number0Elm.classes()).toContain('ERFILTER-ERROR')
-    expect(number1Elm.classes()).toContain('ERFILTER-ERROR')
-    number0Elm.find('input').element.value = '10'
-    await number0Elm.find('input').trigger('input')
-    number1Elm.find('input').element.value = '20'
-    await number1Elm.find('input').trigger('input')
+    selectElm = wrapper.find(utils.getTestId(`${NAME.PICKERCOMPONENT}-0-0`)).find(utils.getTestId(`${NAME.TEXTTYPE}-select`, 'id'))
+    expect(selectElm.classes()).toContain('ERFILTER-ERROR')
+    const inputWrapper = selectElm.find('.el-select__input')
+    const selectInputEl = inputWrapper.element
+    selectInputEl.value = '1'
+    await new Promise(resolve => setTimeout(resolve, 100))
+    await inputWrapper.trigger('input')
+    getSelectOptions(utils.getTestId(`${NAME.TEXTTYPE}-select-popperClass`, 'id'))[0].click()
+    await nextTick()
     expect(wrapper.findComponent({ ref: 'ERfilterRef' }).vm.getData()).toMatchSnapshot()
   })
 })
